fix!: remove /api/post
The code handling /api/post is similar to the code handling /post, but
/api/post has some bugs such as allowing banned users to upload files.
Making /post the only API able to upload files would reduce duplicated
code which reduces the chance of bugs.

Although this removes /api/post, it shouldn't cause any issues since
/api/post isn't currently usable because it requires a multipart
request, but the code uses get_json which only works when the
mimetype is application/json.